+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added custom "compact" style to `DateTime.getWeekdays`.
 - Handling other DST shifts.
 - Added `normalize` flag to `Interval.getCalendarMonth` to always span 6 weeks.
+- Handle setting components in unintended order.
 - Better documentation.
 
 ## 0.3.0